refactor : create config package
