+++ tsconfig.json
@@ -19,6 +19,7 @@
     "paths": {
       "@apis/*": ["src/apis/*"],
       "@assets/*": ["src/assets/*"],
+      "@configs/*": ["src/configs/*"],
       "@models/*": ["src/models/*"],
       "@pages/*": ["src/pages/*"],
       "@routers/*": ["src/routers/*"],
